fix(devcontainers-cli): allow yarn to install when `packageManager` not `yarn`
On our dogfood box, we fail to install `@devcontainers/cli` with `yarn`
because our agent directory `/home/coder/coder` contains a
`package.json` with `packageManager` being set to `pnpm`. This change
instead ensures to run the script inside the `$CODER_SCRIPT_DATA_DIR` so
that we don't read a `package.json` and cause things to break.