refactor(coder/agentapi): support terraform provider coder v2.12.0
In https://github.com/coder/terraform-provider-coder v2.12.0-pre0 we
have removed the requirement for the "AI Prompt" parameter, and are
intending on slightly re-designing the API of the AI task modules.

Instead of `agentapi` defining the `coder_ai_task` resource, it will
output the `task_app_id`. Consumers of the module will then be expected
to create the `coder_ai_task` resource themselves with this `task_app_id`.